Bug fixes in duplicating and removing file duplicates

+++ les01.py
@@ -5,8 +5,8 @@ import sys
 
 print("hello")
 
+is_continue = "y"
 while True:
-    is_continue = raw_input("continue (y/n)? ")
     if is_continue == "y":
 
         print("I can do following: ")
@@ -18,8 +18,10 @@ while True:
         print(" [6] - Remove duplicates")
 
         do = int(raw_input("select, please: "))
+
         if do == 1: # Print file list
             print(os.listdir("."))
+
         elif do == 2: # Print system info
             print("SYSTEM INFO")
             print("platform", sys.platform)
@@ -28,15 +30,17 @@ while True:
             print("current user", os.getlogin())
             print("file system coding", sys.getfilesystemencoding())
             print("CPU count: ", psutil.cpu_count(logical=False))
+
         elif do == 3: # Print process list
             print(psutil.pids())
+
         elif do == 4: #Duplicate file list
             file_list = os.listdir(".")
             print("files in dir:", file_list)
             print("duplicating....")
             i = 0
             while i < len(file_list):
-                if file_list[i][0] == '.':
+                if not os.path.isfile(file_list[i]):
                     i += 1
                     continue
                 print("copying...: ", file_list[i])
@@ -44,6 +48,7 @@ while True:
                 i += 1
             file_list = os.listdir(".")
             print("files in dir:", file_list)
+
         elif do == 5: # Duplicate specified file
             file_list = os.listdir(".")
             print("files in dir:", file_list)
@@ -51,11 +56,15 @@ while True:
             print("duplicating....")
             i = 0
             while i < len(file_list):
+                if not os.path.isfile(file_list[i]):
+                    i += 1
+                    continue
                 if file_list[i] == file_to_dup:
                     shutil.copy(file_list[i], file_list[i] + '.dupl')
                 i += 1
             file_list = os.listdir(".")
             print("files in dir:", file_list)
+
         elif do == 6: # Remove duplicates
             file_list = os.listdir(".")
             print("files in dir:", file_list)
@@ -64,12 +73,14 @@ while True:
             while i < len(file_list):
                 file_name = file_list[i]
                 extension = os.path.splitext(file_name)[1]
+                #if file_name.endswith('.dupl'): # option 2
                 if extension == '.dupl':
                     print("to be removed: ", file_name)
                     os.remove(file_name)
                 i += 1
             file_list = os.listdir(".")
             print("files in dir:", file_list)
+
         else:
             print("unknown operation")
     elif is_continue == "n":
@@ -77,3 +88,5 @@ while True:
         break
     else:
         print("hwat??")
+
+    is_continue = raw_input("continue (y/n)? ")
